Moved NewTestServer to client pkg where it is used.
